Refactor toolbox structure and update imports
- Removed the old toolboxGrade9.js file and replaced it with a new toolbox.js file.
- Updated ToolboxConfigManager to import from the new toolbox.js instead of toolboxGrade9.js.
- Adjusted WorkspaceManager comments and references to reflect the new toolbox structure.
